chore: create new component tree
