+++ packages/pluggableWidgets/datagrid-dropdown-filter-web/src/DatagridDropdownFilter.tsx
@@ -1,42 +1,48 @@
 import { createElement, ReactElement } from "react";
 import { withPreloader } from "@mendix/widget-plugin-platform/hoc/withPreloader";
 import { DatagridDropdownFilterContainerProps } from "../typings/DatagridDropdownFilterProps";
-import { StaticFilterContainer } from "./components/StaticFilterContainer";
-import { withSelectFilterAPI, Select_FilterAPIv2 } from "./hocs/withSelectFilterAPI";
-import { RefFilterContainer } from "./components/RefFilterContainer";
+import { AttrFilter } from "./components/AttrFilter";
 
-function Container(props: DatagridDropdownFilterContainerProps & Select_FilterAPIv2): React.ReactElement {
-    const commonProps = {
-        ariaLabel: props.ariaLabel?.value,
-        className: props.class,
-        tabIndex: props.tabIndex,
-        styles: props.style,
-        onChange: props.onChange,
-        valueAttribute: props.valueAttribute,
-        parentChannelName: props.parentChannelName,
-        name: props.name,
-        multiselect: props.multiSelect,
-        emptyCaption: props.emptyOptionCaption?.value,
-        defaultValue: props.defaultValue?.value,
-        filterable: props.filterable,
-        selectionMethod: props.selectionMethod,
-        selectedItemsStyle: props.selectedItemsStyle,
-        clearable: props.clearable
-    };
+// function Container(props: DatagridDropdownFilterContainerProps & Select_FilterAPIv2): React.ReactElement {
+//     const commonProps = {
+//         ariaLabel: props.ariaLabel?.value,
+//         className: props.class,
+//         tabIndex: props.tabIndex,
+//         styles: props.style,
+//         onChange: props.onChange,
+//         valueAttribute: props.valueAttribute,
+//         parentChannelName: props.parentChannelName,
+//         name: props.name,
+//         multiselect: props.multiSelect,
+//         emptyCaption: props.emptyOptionCaption?.value,
+//         defaultValue: props.defaultValue?.value,
+//         filterable: props.filterable,
+//         selectionMethod: props.selectionMethod,
+//         selectedItemsStyle: props.selectedItemsStyle,
+//         clearable: props.clearable
+//     };
 
-    if (props.filterStore.storeType === "refselect") {
-        return <RefFilterContainer {...commonProps} filterStore={props.filterStore} />;
+//     if (props.filterStore.storeType === "refselect") {
+//         return <RefFilterContainer {...commonProps} filterStore={props.filterStore} />;
+//     }
+
+//     return (
+//         <StaticFilterContainer {...commonProps} filterStore={props.filterStore} filterOptions={props.filterOptions} />
+//     );
+// }
+
+const DatagridDropdownFilter = withPreloader(Container, props => props.defaultValue?.status === "loading");
+
+export default DatagridDropdownFilter;
+
+function Container(props: DatagridDropdownFilterContainerProps): ReactElement {
+    if (props.baseType === "attr") {
+        return <AttrFilter {...props} />;
     }
 
-    return (
-        <StaticFilterContainer {...commonProps} filterStore={props.filterStore} filterOptions={props.filterOptions} />
-    );
+    return <RefFilter {...props} />;
 }
 
-const container = withPreloader(Container, props => props.defaultValue?.status === "loading");
-
-const Widget = withSelectFilterAPI(container);
-
-export default function DatagridDropdownFilter(props: DatagridDropdownFilterContainerProps): ReactElement {
-    return <Widget {...props} />;
+function RefFilter(props: DatagridDropdownFilterContainerProps): ReactElement {
+    return <div key={props.name} />;
 }
